+++ src/main.ts
@@ -3,12 +3,13 @@ import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
 import {environment} from './environments/environment';
 import {AppModule} from './app/app.module';
 import {hmrBootstrap} from './hmr';
+import {NgModuleRef} from '@angular/core/src/linker/ng_module_factory';
 
 if (environment.production) {
 	enableProdMode();
 }
 
-const bootstrap = () => platformBrowserDynamic()
+const bootstrap: any = (): Promise<NgModuleRef<any>> => platformBrowserDynamic()
 	.bootstrapModule(AppModule);
 
 if (environment.hmr) {