+++ CHANGELOG.md
@@ -13,6 +13,8 @@ On top of [Keep a Changelog](http://keepachangelog.com/), you can use the type `
 It should be more verbose to avoid painful experiences and listed in top.
 
 ## [Unreleased]
+
+## [0.2.0]
 ### Added
 - Enhanced configuration for the index.html file
 - Favicons